OPENAI/requirements.txt
Flask  
requests  
flask-cors  
azure-search-documents  
azure-core  
azure-storage-blob  
python-docx  
PyMuPDF